initial BindingCollector
